Seperate solveTime functions which are related to solves
